+++ src/main/java/io/github/kydzombie/crimsonforest/recipe/BasinRecipe.java
@@ -4,5 +4,5 @@ import io.github.kydzombie.crimsonforest.magic.EssenceType;
 import net.minecraft.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
-public record BasinRecipe(@NotNull ItemStack input, @NotNull EssenceType essenceType, int essence, @NotNull ItemStack output) {
+public record BasinRecipe(@NotNull ItemStack input, @NotNull EssenceType essenceType, long essence, @NotNull ItemStack output) {
 }
